requirements.txt
langchain-core
langchain_community 
langgraph 
langchain-groq 
langchain_google_genai
langchain_chroma
chromadb
beautifulsoup4
python-dotenv
colorama
langserve
sse_starlette
uvicorn
gunicorn
fastapi
pytest